backend/requirements.txt
alembic==1.18.4
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
asttokens==3.0.0
asyncpg==0.31.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.2
executing==2.2.0
fastapi==0.135.3
Flask==3.1.3
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.70.0
googleapis-common-protos==1.74.0
googlemaps==4.10.0
greenlet==3.3.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
parso==0.8.4
passlib==1.7.4
pefile==2023.2.7
pillow==12.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytweening==1.2.0
pywhatkit==5.4
pywin32==311
pywin32-ctypes==0.2.3
pyzmq==27.0.0
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==1.0.0
tenacity==9.1.4
tornado==6.5.1
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
wcwidth==0.2.13
websockets==16.0
Werkzeug==3.1.6
wikipedia==1.4.0
